Avanzo #25. Aplico el principio de responsabilidad única
